refactoring connection handler
