+++ server/src/main/java/com/costa/socket/server/service/TaskHandler.java
@@ -3,7 +3,7 @@ package com.costa.socket.server.service;
 import com.costa.socket.server.dto.ServerResponse;
 import com.costa.socket.server.model.CommandType;
 import com.costa.socket.server.model.ResultStatus;
-import com.costa.socket.server.model.UserRequest;
+import com.costa.socket.server.dto.UserRequest;
 import com.costa.socket.server.service.impl.TaskHandlerFactory;
 
 /**
