+++ server/src/main/java/com/costa/socket/server/service/impl/TaskModifyHandler.java
@@ -3,7 +3,7 @@ package com.costa.socket.server.service.impl;
 import com.costa.socket.server.dto.ServerResponse;
 import com.costa.socket.server.mapper.TaskMapper;
 import com.costa.socket.server.model.ResultStatus;
-import com.costa.socket.server.model.UserRequest;
+import com.costa.socket.server.dto.UserRequest;
 import com.costa.socket.server.service.Validator;
 import com.costa.socket.server.service.TaskHandler;
 import com.costa.socket.server.daoservice.TaskService;
